vim: Switch to deoplete for completions (finally)
The only reason why neocomplete stuck around for so long in my ~/.vimrc
was because of compatibility with older Vim installations on servers
that I didn't control.

However, Ubuntu 16.04 LTS plans on ending Standard Support in April
2021, which is the last Ubuntu version that doesn't have Vim 8 out of
the box. Since I'm exclusively running 18.04 LTS or later nowadays, or
have root access to install a more modern Vim via vim-nox/a package that
has Python plugin support from PPAs (or can compile from source if
necessary), it seems safe to switch over to using deoplete.nvim, which
requires Python 3 and Vim 8.

The following alternatives were considered (briefly) before being
discarded:

    * LanguageClient-neovim: Seems like this requires LSP support, and I
      want completions with and without LSP
    * coc.nvim: Requires Node.js, which is a separate runtime. All
      machines I work on already have Python, and installing packages to
      the home directory via 'pip install --user' is good enough for
      deoplete.nvim (although it might be worth exploring Dockerizing)
    * YouCompleteMe: I was never really a fan of how large
      YouCompleteMe was.

We can't pin a release for now since 5.3 doesn't yet exist (which fixes
msgpack incompatibility breakages in 1.0.0). Otherwise, we would need to
install an older version of msgpack for use with 5.2 (which is currently
the latest release). Once the release is cut, I'll update the ~/.vimrc
to use a tag. We _could_ pin a commit, but it seems fine for now as we
move towards pinning versions of every plugin).