atom: Remove atom from dotfiles
I no longer use Atom, so it doesn't seem to make sense to keep a config
file for a program that won't be used on my computer. There's also no
guarantee that the config file actually works on the latest version of
Atom, since I haven't opened the editor in a long time.