dircolors: Work around C_MISSING bug
There seems to be a bug either in readline or bash-completion somewhere,
where bash-completion makes readline think that it is completing file
names (which is not true when suggesting options). This results in the
options being interpreted as missing files. There also seems to be an
inconsistency between the behaviour of coreutils and readline: readline
seems to use the C_MISSING color code to highlight files that do not
exist, while it appears that coreutils uses this to highlight broken
symlink targets.

As a workaround, remove the flashing option from C_MISSING and red
background colour.