git: Align init.defaultBranch with GitHub defaults
The default branch name for new repositories created on GitHub is now
'main'. This aligns with initiatives at work as alternative names for
the default branch name.

As of Git 2.28, the init.defaultBranch configuration option is supported
to specify the default branch name in newly created Git repositories via
'git init'. Since Git just treats options in this file as key-value
pairs, this should just be ignored on older versions of Git.